+++ types/types.ts
@@ -1,0 +1,7 @@
+export { User } from "./user";
+export { Send } from "./send";
+export { Route } from "./route";
+export { Tag } from "./tag";
+export { Forum } from "./forum";
+export { Post } from "./post";
+export { Comment } from "./comment";
